updated udp:os_send to use unsigned short

+++ src/libocean/libocean.h
@@ -25,7 +25,7 @@ char * os_meta_create (int, char[]);
 
 void os_meta_dump (void *);
 
-int os_send (void *, int, const char *);
+int os_send (void *, unsigned short int, const char *);
 
 void os_recv (void);
